+++ src/tribler/core/components/component.py
@@ -10,7 +10,7 @@ from tribler.core.components.exceptions import ComponentStartupException, Missed
 from tribler.core.components.reporter.exception_handler import default_core_exception_handler
 from tribler.core.sentry_reporter.sentry_reporter import SentryReporter
 from tribler.core.utilities.db_corruption_handling.base import DatabaseIsCorrupted
-from tribler.core.utilities.exit_codes import EXITCODE_DATABASE_IS_CORRUPTED
+from tribler.core.utilities.exit_codes.tribler_exit_codes import EXITCODE_DATABASE_IS_CORRUPTED
 from tribler.core.utilities.process_manager import get_global_process_manager
 
 if TYPE_CHECKING:
