+++ src/tribler/core/components/reporter/exception_handler.py
@@ -11,7 +11,7 @@ from tribler.core.components.exceptions import ComponentStartupException
 from tribler.core.components.reporter.reported_error import ReportedError
 from tribler.core.sentry_reporter.sentry_reporter import SentryReporter
 from tribler.core.utilities.db_corruption_handling.base import DatabaseIsCorrupted
-from tribler.core.utilities.exit_codes import EXITCODE_DATABASE_IS_CORRUPTED
+from tribler.core.utilities.exit_codes.tribler_exit_codes import EXITCODE_DATABASE_IS_CORRUPTED
 from tribler.core.utilities.process_manager import get_global_process_manager
 
 # There are some errors that we are ignoring.
